move utility to `deepmd_utils` (without modifaction) (#3140)
Move framework-independent codes to the `deepmd_utils` module without
modification, as a step of
https://github.com/deepmodeling/deepmd-kit/issues/3118.

---------
