+++ deepmd_utils/loggers/__init__.py
@@ -1,0 +1,8 @@
+# SPDX-License-Identifier: LGPL-3.0-or-later
+"""Module taking care of logging duties."""
+
+from .loggers import (
+    set_log_handles,
+)
+
+__all__ = ["set_log_handles"]
