+++ deepmd_utils/utils/__init__.py
@@ -1,0 +1,3 @@
+# SPDX-License-Identifier: LGPL-3.0-or-later
+# For performance, do not add things to this file
+# import submodules instead
